# Jordan-AI-Platform/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  label: string;
  labelEn: string;
  description: string;
  count: number;
  accent: string;
}

export type SourceType = typeof SourceType[keyof typeof SourceType];


export const SourceType = {
  official: 'official',
  authoritative: 'authoritative',
  reliable: 'reliable',
  secondary: 'secondary',
} as const;

export interface Source {
  id: string;
  title: string;
  publisher: string;
  domain: string;
  url: string;
  type: SourceType;
  retrievedAt: string;
  /** @nullable */
  publicationDate: string | null;
}

export type KnowledgeItemStatus = typeof KnowledgeItemStatus[keyof typeof KnowledgeItemStatus];


export const KnowledgeItemStatus = {
  verified: 'verified',
  review: 'review',
} as const;

export interface KnowledgeItem {
  id: string;
  title: string;
  titleEn: string;
  summary: string;
  category: string;
  categoryLabel: string;
  status: KnowledgeItemStatus;
  /** @nullable */
  lastVerified: string | null;
  source: Source;
  tags: string[];
}

export interface SearchResponse {
  query: string;
  searchedLive: boolean;
  message: string;
  results: KnowledgeItem[];
}

export interface HomeSummary {
  verifiedCount: number;
  categoryCount: number;
  sourceCount: number;
  lastUpdated: string;
  categories: Category[];
  featured: KnowledgeItem[];
  popularQuestions: string[];
}

export type AnswerQuestionInputLanguage = typeof AnswerQuestionInputLanguage[keyof typeof AnswerQuestionInputLanguage];


export const AnswerQuestionInputLanguage = {
  ar: 'ar',
  en: 'en',
} as const;

export interface AnswerQuestionInput {
  /**
     * @minLength 2
     * @maxLength 1000
     */
  question: string;
  language?: AnswerQuestionInputLanguage;
}

export type AssistantAnswerStatus = typeof AssistantAnswerStatus[keyof typeof AssistantAnswerStatus];


export const AssistantAnswerStatus = {
  verified: 'verified',
  'needs-current-source': 'needs-current-source',
  'not-verified': 'not-verified',
} as const;

export interface AssistantAnswer {
  question: string;
  answer: string;
  status: AssistantAnswerStatus;
  searchedLive: boolean;
  sources: Source[];
  note: string;
}

export type FeedbackInputKind = typeof FeedbackInputKind[keyof typeof FeedbackInputKind];


export const FeedbackInputKind = {
  incorrect: 'incorrect',
  outdated: 'outdated',
  missing: 'missing',
  'bad-source': 'bad-source',
  'answer-problem': 'answer-problem',
} as const;

export interface FeedbackInput {
  kind: FeedbackInputKind;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  message: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  knowledgeId?: string | null;
}

export interface Feedback {
  id: string;
  received: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type SearchJordanParams = {
/**
 * @minLength 1
 */
q: string;
category?: string;
language?: SearchJordanLanguage;
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type SearchJordanLanguage = typeof SearchJordanLanguage[keyof typeof SearchJordanLanguage];


export const SearchJordanLanguage = {
  ar: 'ar',
  en: 'en',
} as const;

export type ListKnowledgeParams = {
category?: string;
/**
 * @minimum 1
 * @maximum 50
 */
limit?: number;
};