# Jordan-AI-Platform/lib/api-zod/src/generated/types/answerQuestionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { AnswerQuestionInputLanguage } from './answerQuestionInputLanguage';

export interface AnswerQuestionInput {
  /**
     * @minLength 2
     * @maxLength 1000
     */
  question: string;
  language?: AnswerQuestionInputLanguage;
}