# Jordan-AI-Platform/lib/api-zod/src/generated/types/assistantAnswer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { AssistantAnswerStatus } from './assistantAnswerStatus';
import type { Source } from './source';

export interface AssistantAnswer {
  question: string;
  answer: string;
  status: AssistantAnswerStatus;
  searchedLive: boolean;
  sources: Source[];
  note: string;
}