# Jordan-AI-Platform/lib/api-zod/src/generated/types/category.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */

export interface Category {
  id: string;
  label: string;
  labelEn: string;
  description: string;
  count: number;
  accent: string;
}