# Jordan-AI-Platform/lib/api-zod/src/generated/types/feedback.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */

export interface Feedback {
  id: string;
  received: boolean;
  message: string;
}