# Jordan-AI-Platform/lib/api-zod/src/generated/types/feedbackInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackInputKind } from './feedbackInputKind';

export interface FeedbackInput {
  kind: FeedbackInputKind;
  /**
     * @minLength 3
     * @maxLength 2000
     */
  message: string;
  /** @nullable */
  question?: string | null;
  /** @nullable */
  knowledgeId?: string | null;
}