# Jordan-AI-Platform/lib/api-zod/src/generated/types/homeSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { Category } from './category';
import type { KnowledgeItem } from './knowledgeItem';

export interface HomeSummary {
  verifiedCount: number;
  categoryCount: number;
  sourceCount: number;
  lastUpdated: string;
  categories: Category[];
  featured: KnowledgeItem[];
  popularQuestions: string[];
}