# Jordan-AI-Platform/lib/api-zod/src/generated/types/knowledgeItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { KnowledgeItemStatus } from './knowledgeItemStatus';
import type { Source } from './source';

export interface KnowledgeItem {
  id: string;
  title: string;
  titleEn: string;
  summary: string;
  category: string;
  categoryLabel: string;
  status: KnowledgeItemStatus;
  /** @nullable */
  lastVerified: string | null;
  source: Source;
  tags: string[];
}