# Jordan-AI-Platform/lib/api-zod/src/generated/types/source.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for the Jordan AI knowledge platform
 * OpenAPI spec version: 0.1.0
 */
import type { SourceType } from './sourceType';

export interface Source {
  id: string;
  title: string;
  publisher: string;
  domain: string;
  url: string;
  type: SourceType;
  retrievedAt: string;
  /** @nullable */
  publicationDate: string | null;
}